fix(devin-lang): simplify regex patterns and add support for Java code blocks 101
The commit message should explain the changes made to the regex patterns in the DevInParser.bnf file, as well as the addition of support for Java code blocks in the test data. The message should also mention that the DevInCompletionContributor has been updated to support the new syntax, and that the DevInParserDefinition now uses a factory method to create elements. The message should also highlight the creation of a new test file that contains a Java code block.