feat(compiler): add support for committing changes
This commit adds support for committing changes in the DevIns compiler. It introduces a new `CommitInsCommand` class that executes the commit operation using `LocalChangesCommitter`. The commit message is passed as a parameter to the `CommitInsCommand`. Additionally, the `BuiltinCommandProvider` now has a new property `requireProps`, which indicates if a command requires additional properties.