fix(git): fix 222 & 233 version GitUtil class for committing local changes #101
This commit adds a new class `GitUtil` in the `cc.unitmesh.devti.language.git` package, which provides a `doCommit` function. This function takes a project, a local change list, and a commit message as parameters, and commits the changes using the IntelliJ VCS API. This new class will be used to commit local changes in the `CommitInsCommand` class. Additionally, the `CommitInsCommand` class is modified to use the `GitUtil` class for performing commits. The commit also includes minor changes to the `README.md` file.

Refs #233